+++ CHANGELOG.md
@@ -1,12 +1,26 @@
 ## lifebit-ai/cloudos-cli: changelog
 
-## v2.35.0 (2025-06-30)
+## v2.37.0 (2025-06-30)
 
 ### Feat
 
 - Allows **specifying** different projects for parameter options i.e. `-p --file1=PROJECT1/Data/input.csv -p --file2=Project2/Data/input.csv` 
 
-## v2.34.0 (2025-06-27)
+## v2.36.0 (2025-06-27)
+
+### Feat
+
+- Adds command to remove files.
+- Adds ci tests for cp and mkdir
+- Patch s3 folders listing files
+
+## v2.35.0 (2025-06-27)
+
+### Feat
+
+-  Adds command to create new folders
+
+## v2.34.0 (2025-06-25)
 
 ### Feat